Fix pulling in the global __INITIAL_STATE__
The events data was being requested on load even though SSR had placed
it in the `window.__INITIAL_STATE__`, because we needed to reference
`globalThis` instead of `this`. In order to do that eslint needed to be
upgraded and the environment set to `es2020: true`. This should be fine
since `globalThis` is supported for at least 4 browser versions in all
major browsers.